Add a form for user to create password
This form asks the user to enter a password and also retype the password in order to confirm. The form will only be submitted if the password and also the confirmed password they entered are the same.
I used web3.js to create an account for the user. After the account is created, the user will be redirected to another page, where their public address and private key will be displayed.